requirements.txt
alembic==1.7.7
amqp==5.1.1
async-timeout==4.0.2
Babel==2.10.3
billiard==3.6.4.0
blinker==1.5
celery==5.2.7
certifi==2022.6.15
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
dnspython==2.2.1
elastic-transport==8.1.2
elasticsearch==8.3.3
email-validator==1.2.1
Flask==2.1.2
Flask-Babel==2.0.0
Flask-Login==0.6.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Moment==1.0.4
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
Mako==1.2.0
MarkupSafe==2.1.1
packaging==21.3
prompt-toolkit==3.0.30
psycopg2==2.9.3
PyJWT==2.4.0
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.2.1
redis==4.3.4
six==1.16.0
SQLAlchemy==1.4.36
urllib3==1.26.11
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.1.2
wrapt==1.14.1
WTForms==3.0.1
zipp==3.8.0
